Refactor: CanvasTab -> PlotDisplay
